pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-fintech"
version = "0.1.0"
description = "RAG FinTech"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Xiaojing Liu", email = "liuxiaojing11@126.com"}
]
keywords = ["rag", "fintech", "document-processing", "nlp"]

dependencies = [
    "fastapi>=0.121.3",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.0.0",
    "PyPDF2>=3.0.1",
    "llama-parse>=0.6.81",
    "llama-index-core>=0.10.0",
    "beautifulsoup4>=4.12.2",
    "markdown-it-py==4.0.0",
    "openai==2.8.1",
    "pymilvus>=2.6.3",
    "SQLAlchemy>=2.0.0",
    "psycopg[binary,pool]==3.3.2",
    "langgraph-checkpoint-postgres>=3.0.4",
    "neo4j==6.0.3",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "redis==7.1.0",
    "hiredis==3.3.0",
    "ragas==0.3.9",
    "datasets==4.4.1",
    "tqdm>=4.67.1",
    "transformers==4.57.3",
    "langchain>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "langchain-deepseek>=1.0.1",
    "langsmith>=0.6.0",
    "rq==2.6.1",
    "tenacity>=9.1.2",
]

[project.optional-dependencies]
dev = [
    "ipdb==0.13.13",
    "debugpy>=1.8.0",
    "ipykernel>=6.25.0",
    "black>=24.0.0",
]

[project.scripts]
rag-api = "api.rag_server:app"
rq-worker = "api.worker:main"

[tool.setuptools]
packages = ["api", "agent", "common", "graphrag", "rag", "repository"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py312"

[tool.black]
line-length = 120
target-version = ['py312'] 
include = '\.pyi?$'
skip-string-normalization = false